+++ Src/Newtonsoft.Json.UnityConverters.Tests/ConvertingUnityTypes/AI/NavMesh/NavMeshHitTests.cs
@@ -2,7 +2,7 @@
 using UnityEngine;
 using UnityEngine.AI;
 
-namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.AI
+namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.AI.NavMesh
 {
     public class NavMeshHitTests : ValueTypeTester<NavMeshHit>
     {
